fix: Amend nav to correct error
Resolves: DCMAW-10312
